refactor: move algorithm registry into `core/registry.py`
